app/requirements.txt
flask==1.0.2
python-dotenv==0.10.1
flask-wtf==0.14.3
pytz==2020.1
azure-cosmos==3.0.2
requests==2.21.0
flask-mail==0.9.1
sendgrid==5.6.0
razorpay==0.2.0
redis==2.10.6
Flask-Session==0.3.1
Flask-Compress==1.4.0
Flask-HTMLmin==1.5.0
mastercard-oauth1-signer
azure-storage-blob==1.5.0
azure-cosmosdb-table==1.0.6
samesite-compat-check==0.2.0
httpagentparser==1.9.0
itsdangerous==1.1.0
MarkupSafe==2.1.1
Jinja2==3.0.0
pymsteams==0.2.1
Werkzeug==2.0.0
httpx==0.23.0
Flask-SocketIO==5.3.1
Pillow==9.2.0
pandas==1.5.3
numpy==1.24.2